Renamed apply_stereo_rectification.py->apply_sr.py and stereo_rectification_map_gen.py->sr_map_gen.py
